Wavesurfer Edits
Added more songs and styling
