+++ TestPassport/public/register.html
@@ -1,0 +1,27 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+  <title>Passport Authentication</title>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1">
+</head>
+
+<body>
+
+<h1>Register</h1>
+<form action="/api/register" method="POST">
+    <div>
+        <label for = "username">Username</label>
+        <input type = "text" id="username" name="username" required>
+    </div>
+    <div>
+        <label for = "password">Password</label>
+        <input type = "text" id="password" name="password" required>
+    </div>
+    <button type="submit">Register</button>
+</form>
+<a href="/login">Login</a>
+
+</body>
+</html>
