--- login/public/index.html
@@ -1,38 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-  <title>Passport Authentication</title>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1">
-</head>
-
-<body>
-
-<h1>Welcome</h1>
-<form action="/logout?_method=DELETE" method="POST">
-    <button type="submit">Log Out</button>
-</form>
-<div id="waveform"></div>
-<button type="submit" id="play">Play/Pause</button>
-
-</body>
-
-<script src="https://unpkg.com/wavesurfer.js"></script>
-
-<script>
-    var wavesurfer = WaveSurfer.create({
-    container: '#waveform',
-    waveColor: 'violet',
-    progressColor: 'purple'
-    });
-
-    wavesurfer.load('./assets/Smack_That.mp3');
-
-    wavesurfer.on('ready', function () {
-    wavesurfer.play();
-});
-
-document.querySelector('[id="play"]').addEventListener('click', wavesurfer.playPause.bind(wavesurfer))
-</script>
-</html>
